1. Move from System.IO.MemoryMappedFiles back to using the NativePlatform class.
2. Add platform specific build configurations for x86-Windows,
x64-Windows, x86-Linux, and x64-Linux.

On Windows, System.IO.MemoryMappedFiles.MemoryMappedFile.CreateFromFile
was unable to create a mapping for the currently running executable. It
did work fine under Mono on Linux, however.

To fix this we switched to calling the Win32 apis directly.

A posix implementation of NativePlatform has not been implemented yet.