Checkin current work.
1. Strings read from the string heap are now cached in the types that
encaspulate the metadata record containing the string index rather than in
the PEFile class. This is because it is possible (particularly with large
metadata files) to have heap indexes appear at unaligned addresses.
Its not as convienient as keeping the cache in the PF file, but it avoids
the need to do repeated dereferences of the same unaligned pointer.

2. Consequently the implementation of PEFile.ReadSystemString is much
simpiler.

3. Fixes up a few problems with validation of the Metadata table header.

4. Adds code to compute the sizes of coded indexes.

5. Adds some more validation to the Assembly constructor

6. Implements ModuleCollection.LoadModule.