Make conversions between 1-based and 0-based indexes explicit in the type system.
This is done by adding ZeroBasedIndex and OneBasedIndex structs, and then
defining explicit type conversions between them.

I had serveral off-by-1 bugs related to the fact that we need to convert
1-based metadata foreign key indexes into the 0-based indexes in the
public facing API. By eliciting the help of the type system it should be
more difficult for me to make these kinds of mistakes.